Integrate move vault logic form deprecated config.ts to new config.schema.ts
